Module aws_sdk_pinpointemail::error

Modules

account_suspended_exception

See AccountSuspendedException

already_exists_exception

See AlreadyExistsException

bad_request_exception

See BadRequestException

concurrent_modification_exception

See ConcurrentModificationException

limit_exceeded_exception

See LimitExceededException

mail_from_domain_not_verified_exception

See MailFromDomainNotVerifiedException

message_rejected

See MessageRejected

not_found_exception

See NotFoundException

sending_paused_exception

See SendingPausedException

too_many_requests_exception

See TooManyRequestsException

Structs

AccountSuspendedException

The message can't be sent because the account's ability to send email has been

AlreadyExistsException

The resource specified in your request already exists.

BadRequestException

The input you provided is invalid.

ConcurrentModificationException

The resource is being modified by another operation or thread.

CreateConfigurationSetError

CreateConfigurationSetEventDestinationError

CreateDedicatedIpPoolError

CreateDeliverabilityTestReportError

CreateEmailIdentityError

DeleteConfigurationSetError

DeleteConfigurationSetEventDestinationError

DeleteDedicatedIpPoolError

DeleteEmailIdentityError

GetAccountError

GetBlacklistReportsError

GetConfigurationSetError

GetConfigurationSetEventDestinationsError

GetDedicatedIpError

GetDedicatedIpsError

GetDeliverabilityDashboardOptionsError

GetDeliverabilityTestReportError

GetDomainDeliverabilityCampaignError

GetDomainStatisticsReportError

GetEmailIdentityError

LimitExceededException

There are too many instances of the specified resource type.

ListConfigurationSetsError

ListDedicatedIpPoolsError

ListDeliverabilityTestReportsError

ListDomainDeliverabilityCampaignsError

ListEmailIdentitiesError

ListTagsForResourceError

MailFromDomainNotVerifiedException

The message can't be sent because the sending domain isn't verified.

MessageRejected

The message can't be sent because it contains invalid content.

NotFoundException

The resource you attempted to access doesn't exist.

PutAccountDedicatedIpWarmupAttributesError

PutAccountSendingAttributesError

PutConfigurationSetDeliveryOptionsError

PutConfigurationSetReputationOptionsError

PutConfigurationSetSendingOptionsError

PutConfigurationSetTrackingOptionsError

PutDedicatedIpInPoolError

PutDedicatedIpWarmupAttributesError

PutDeliverabilityDashboardOptionError

PutEmailIdentityDkimAttributesError

PutEmailIdentityFeedbackAttributesError

PutEmailIdentityMailFromAttributesError

SendEmailError

SendingPausedException

The message can't be sent because the account's ability to send email is currently

TagResourceError

TooManyRequestsException

Too many requests have been made to the operation.

UntagResourceError

UpdateConfigurationSetEventDestinationError

Enums

CreateConfigurationSetErrorKind

CreateConfigurationSetEventDestinationErrorKind

CreateDedicatedIpPoolErrorKind

CreateDeliverabilityTestReportErrorKind

CreateEmailIdentityErrorKind

DeleteConfigurationSetErrorKind

DeleteConfigurationSetEventDestinationErrorKind

DeleteDedicatedIpPoolErrorKind

DeleteEmailIdentityErrorKind

GetAccountErrorKind

GetBlacklistReportsErrorKind

GetConfigurationSetErrorKind

GetConfigurationSetEventDestinationsErrorKind

GetDedicatedIpErrorKind

GetDedicatedIpsErrorKind

GetDeliverabilityDashboardOptionsErrorKind

GetDeliverabilityTestReportErrorKind

GetDomainDeliverabilityCampaignErrorKind

GetDomainStatisticsReportErrorKind

GetEmailIdentityErrorKind

ListConfigurationSetsErrorKind

ListDedicatedIpPoolsErrorKind

ListDeliverabilityTestReportsErrorKind

ListDomainDeliverabilityCampaignsErrorKind

ListEmailIdentitiesErrorKind

ListTagsForResourceErrorKind

PutAccountDedicatedIpWarmupAttributesErrorKind

PutAccountSendingAttributesErrorKind

PutConfigurationSetDeliveryOptionsErrorKind

PutConfigurationSetReputationOptionsErrorKind

PutConfigurationSetSendingOptionsErrorKind

PutConfigurationSetTrackingOptionsErrorKind

PutDedicatedIpInPoolErrorKind

PutDedicatedIpWarmupAttributesErrorKind

PutDeliverabilityDashboardOptionErrorKind

PutEmailIdentityDkimAttributesErrorKind

PutEmailIdentityFeedbackAttributesErrorKind

PutEmailIdentityMailFromAttributesErrorKind

SendEmailErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateConfigurationSetEventDestinationErrorKind