Module aws_sdk_pinpointemail::input

Modules

create_configuration_set_event_destination_input

See CreateConfigurationSetEventDestinationInput

create_configuration_set_input

See CreateConfigurationSetInput

create_dedicated_ip_pool_input

See CreateDedicatedIpPoolInput

create_deliverability_test_report_input

See CreateDeliverabilityTestReportInput

create_email_identity_input

See CreateEmailIdentityInput

delete_configuration_set_event_destination_input

See DeleteConfigurationSetEventDestinationInput

delete_configuration_set_input

See DeleteConfigurationSetInput

delete_dedicated_ip_pool_input

See DeleteDedicatedIpPoolInput

delete_email_identity_input

See DeleteEmailIdentityInput

get_account_input

See GetAccountInput

get_blacklist_reports_input

See GetBlacklistReportsInput

get_configuration_set_event_destinations_input

See GetConfigurationSetEventDestinationsInput

get_configuration_set_input

See GetConfigurationSetInput

get_dedicated_ip_input

See GetDedicatedIpInput

get_dedicated_ips_input

See GetDedicatedIpsInput

get_deliverability_dashboard_options_input

See GetDeliverabilityDashboardOptionsInput

get_deliverability_test_report_input

See GetDeliverabilityTestReportInput

get_domain_deliverability_campaign_input

See GetDomainDeliverabilityCampaignInput

get_domain_statistics_report_input

See GetDomainStatisticsReportInput

get_email_identity_input

See GetEmailIdentityInput

list_configuration_sets_input

See ListConfigurationSetsInput

list_dedicated_ip_pools_input

See ListDedicatedIpPoolsInput

list_deliverability_test_reports_input

See ListDeliverabilityTestReportsInput

list_domain_deliverability_campaigns_input

See ListDomainDeliverabilityCampaignsInput

list_email_identities_input

See ListEmailIdentitiesInput

list_tags_for_resource_input

See ListTagsForResourceInput

put_account_dedicated_ip_warmup_attributes_input

See PutAccountDedicatedIpWarmupAttributesInput

put_account_sending_attributes_input

See PutAccountSendingAttributesInput

put_configuration_set_delivery_options_input

See PutConfigurationSetDeliveryOptionsInput

put_configuration_set_reputation_options_input

See PutConfigurationSetReputationOptionsInput

put_configuration_set_sending_options_input

See PutConfigurationSetSendingOptionsInput

put_configuration_set_tracking_options_input

See PutConfigurationSetTrackingOptionsInput

put_dedicated_ip_in_pool_input

See PutDedicatedIpInPoolInput

put_dedicated_ip_warmup_attributes_input

See PutDedicatedIpWarmupAttributesInput

put_deliverability_dashboard_option_input

See PutDeliverabilityDashboardOptionInput

put_email_identity_dkim_attributes_input

See PutEmailIdentityDkimAttributesInput

put_email_identity_feedback_attributes_input

See PutEmailIdentityFeedbackAttributesInput

put_email_identity_mail_from_attributes_input

See PutEmailIdentityMailFromAttributesInput

send_email_input

See SendEmailInput

tag_resource_input

See TagResourceInput

untag_resource_input

See UntagResourceInput

update_configuration_set_event_destination_input

See UpdateConfigurationSetEventDestinationInput

Structs

CreateConfigurationSetEventDestinationInput

A request to add an event destination to a configuration set.

CreateConfigurationSetInput

A request to create a configuration set.

CreateDedicatedIpPoolInput

A request to create a new dedicated IP pool.

CreateDeliverabilityTestReportInput

A request to perform a predictive inbox placement test. Predictive inbox placement tests can help you predict how your messages will

CreateEmailIdentityInput

A request to begin the verification process for an email identity (an email address or

DeleteConfigurationSetEventDestinationInput

A request to delete an event destination from a configuration set.

DeleteConfigurationSetInput

A request to delete a configuration set.

DeleteDedicatedIpPoolInput

A request to delete a dedicated IP pool.

DeleteEmailIdentityInput

A request to delete an existing email identity. When you delete an identity, you lose

GetAccountInput

A request to obtain information about the email-sending capabilities of your Amazon Pinpoint

GetBlacklistReportsInput

A request to retrieve a list of the blacklists that your dedicated IP addresses appear

GetConfigurationSetEventDestinationsInput

A request to obtain information about the event destinations for a configuration

GetConfigurationSetInput

A request to obtain information about a configuration set.

GetDedicatedIpInput

A request to obtain more information about a dedicated IP address.

GetDedicatedIpsInput

A request to obtain more information about dedicated IP pools.

GetDeliverabilityDashboardOptionsInput

Retrieve information about the status of the Deliverability dashboard for your Amazon Pinpoint account.

GetDeliverabilityTestReportInput

A request to retrieve the results of a predictive inbox placement test.

GetDomainDeliverabilityCampaignInput

Retrieve all the deliverability data for a specific campaign. This data is available

GetDomainStatisticsReportInput

A request to obtain deliverability metrics for a domain.

GetEmailIdentityInput

A request to return details about an email identity.

ListConfigurationSetsInput

A request to obtain a list of configuration sets for your Amazon Pinpoint account in the current

ListDedicatedIpPoolsInput

A request to obtain a list of dedicated IP pools.

ListDeliverabilityTestReportsInput

A request to list all of the predictive inbox placement tests that you've performed.

ListDomainDeliverabilityCampaignsInput

Retrieve deliverability data for all the campaigns that used a specific domain to send

ListEmailIdentitiesInput

A request to list all of the email identities associated with your Amazon Pinpoint account. This

ListTagsForResourceInput

PutAccountDedicatedIpWarmupAttributesInput

A request to enable or disable the automatic IP address warm-up feature.

PutAccountSendingAttributesInput

A request to change the ability of your account to send email.

PutConfigurationSetDeliveryOptionsInput

A request to associate a configuration set with a dedicated IP pool.

PutConfigurationSetReputationOptionsInput

A request to enable or disable tracking of reputation metrics for a configuration

PutConfigurationSetSendingOptionsInput

A request to enable or disable the ability of Amazon Pinpoint to send emails that use a specific

PutConfigurationSetTrackingOptionsInput

A request to add a custom domain for tracking open and click events to a configuration

PutDedicatedIpInPoolInput

A request to move a dedicated IP address to a dedicated IP pool.

PutDedicatedIpWarmupAttributesInput

A request to change the warm-up attributes for a dedicated IP address. This operation

PutDeliverabilityDashboardOptionInput

Enable or disable the Deliverability dashboard for your Amazon Pinpoint account. When you enable the

PutEmailIdentityDkimAttributesInput

A request to enable or disable DKIM signing of email that you send from an email

PutEmailIdentityFeedbackAttributesInput

A request to set the attributes that control how bounce and complaint events are

PutEmailIdentityMailFromAttributesInput

A request to configure the custom MAIL FROM domain for a verified identity.

SendEmailInput

A request to send an email message.

TagResourceInput

UntagResourceInput

UpdateConfigurationSetEventDestinationInput

A request to change the settings for an event destination for a configuration