Module aws_sdk_pinpointemail::model

Modules

blacklist_entry

See BlacklistEntry

body

See Body

cloud_watch_destination

See CloudWatchDestination

cloud_watch_dimension_configuration

See CloudWatchDimensionConfiguration

content

See Content

daily_volume

See DailyVolume

dedicated_ip

See DedicatedIp

deliverability_test_report

See DeliverabilityTestReport

delivery_options

See DeliveryOptions

destination

See Destination

dkim_attributes

See DkimAttributes

domain_deliverability_campaign

See DomainDeliverabilityCampaign

domain_deliverability_tracking_option

See DomainDeliverabilityTrackingOption

domain_isp_placement

See DomainIspPlacement

email_content

See EmailContent

event_destination

See EventDestination

event_destination_definition

See EventDestinationDefinition

identity_info

See IdentityInfo

inbox_placement_tracking_option

See InboxPlacementTrackingOption

isp_placement

See IspPlacement

kinesis_firehose_destination

See KinesisFirehoseDestination

mail_from_attributes

See MailFromAttributes

message

See Message

message_tag

See MessageTag

overall_volume

See OverallVolume

pinpoint_destination

See PinpointDestination

placement_statistics

See PlacementStatistics

raw_message

See RawMessage

reputation_options

See ReputationOptions

send_quota

See SendQuota

sending_options

See SendingOptions

sns_destination

See SnsDestination

tag

See Tag

template

See Template

tracking_options

See TrackingOptions

volume_statistics

See VolumeStatistics

Structs

BlacklistEntry

An object that contains information about a blacklisting event that impacts one of the

Body

Represents the body of the email message.

CloudWatchDestination

An object that defines an Amazon CloudWatch destination for email events. You can use Amazon CloudWatch to

CloudWatchDimensionConfiguration

An object that defines the dimension configuration to use when you send Amazon Pinpoint email

Content

An object that represents the content of the email, and optionally a character set

DailyVolume

An object that contains information about the volume of email sent on each day of the

DedicatedIp

Contains information about a dedicated IP address that is associated with your Amazon Pinpoint

DeliverabilityTestReport

An object that contains metadata related to a predictive inbox placement test.

DeliveryOptions

Used to associate a configuration set with a dedicated IP pool.

Destination

An object that describes the recipients for an email.

DkimAttributes

An object that contains information about the DKIM configuration for an email

DomainDeliverabilityCampaign

An object that contains the deliverability data for a specific campaign. This data is

DomainDeliverabilityTrackingOption

An object that contains information about the Deliverability dashboard subscription for a

DomainIspPlacement

An object that contains inbox placement data for email sent from one of your email

EmailContent

An object that defines the entire content of the email, including the message headers

EventDestination

In Amazon Pinpoint, events include message sends, deliveries, opens,

EventDestinationDefinition

An object that defines the event destination. Specifically, it defines which services

IdentityInfo

Information about an email identity.

InboxPlacementTrackingOption

An object that contains information about the inbox placement data settings for a

IspPlacement

An object that describes how email sent during the predictive inbox placement test was handled by a certain

KinesisFirehoseDestination

An object that defines an Amazon Kinesis Data Firehose destination for email events. You can use Amazon Kinesis Data Firehose to

MailFromAttributes

A list of attributes that are associated with a MAIL FROM domain.

Message

Represents the email message that you're sending. The Message object

MessageTag

Contains the name and value of a tag that you apply to an email. You can use message

OverallVolume

An object that contains information about email that was sent from the selected

PinpointDestination

An object that defines a Amazon Pinpoint destination for email events. You can use Amazon Pinpoint events

PlacementStatistics

An object that contains inbox placement data for an email provider.

RawMessage

The raw email message.

ReputationOptions

Enable or disable collection of reputation metrics for emails that you send using this

SendQuota

An object that contains information about the per-day and per-second sending limits

SendingOptions

Used to enable or disable email sending for messages that use this configuration set

SnsDestination

An object that defines an Amazon SNS destination for email events. You can use Amazon SNS to

Tag

An object that defines the tags that are associated with a resource.

Template

TrackingOptions

An object that defines the tracking options for a configuration set. When you use

VolumeStatistics

An object that contains information about the amount of email that was delivered to

Enums

BehaviorOnMxFailure

The action that you want Amazon Pinpoint to take if it can't read the required MX record for a

DeliverabilityDashboardAccountStatus

The current status of your Deliverability dashboard subscription. If this value is

DeliverabilityTestStatus

The status of a predictive inbox placement test. If the status is IN_PROGRESS, then the predictive inbox placement test is

DimensionValueSource

The location where Amazon Pinpoint finds the value of a dimension to publish to Amazon CloudWatch. If you

DkimStatus

The DKIM authentication status of the identity. The status can be one of the

EventType

An email sending event type. For example, email sends, opens, and bounces are all

IdentityType

The email identity type. The identity type can be one of the following:

MailFromDomainStatus

The status of the MAIL FROM domain. This status can have the following values:

TlsPolicy

Specifies whether messages that use the configuration set are required to use

WarmupStatus

The warmup status of a dedicated IP.