Module aws_sdk_pinpointemail::operation

Structs

CreateConfigurationSet

Create a configuration set. Configuration sets are groups of

CreateConfigurationSetEventDestination

Create an event destination. In Amazon Pinpoint, events include message

CreateDedicatedIpPool

Create a new pool of dedicated IP addresses. A pool can include one or more dedicated

CreateDeliverabilityTestReport

Create a new predictive inbox placement test. Predictive inbox placement tests can help you predict how your messages will be handled

CreateEmailIdentity

Verifies an email identity for use with Amazon Pinpoint. In Amazon Pinpoint, an identity is an email

DeleteConfigurationSet

Delete an existing configuration set.

DeleteConfigurationSetEventDestination

Delete an event destination.

DeleteDedicatedIpPool

Delete a dedicated IP pool.

DeleteEmailIdentity

Deletes an email identity that you previously verified for use with Amazon Pinpoint. An identity

GetAccount

Obtain information about the email-sending status and capabilities of your Amazon Pinpoint

GetBlacklistReports

Retrieve a list of the blacklists that your dedicated IP addresses appear on.

GetConfigurationSet

Get information about an existing configuration set, including the dedicated IP pool

GetConfigurationSetEventDestinations

Retrieve a list of event destinations that are associated with a configuration

GetDedicatedIp

Get information about a dedicated IP address, including the name of the dedicated IP

GetDedicatedIps

List the dedicated IP addresses that are associated with your Amazon Pinpoint

GetDeliverabilityDashboardOptions

Retrieve information about the status of the Deliverability dashboard for your Amazon Pinpoint account.

GetDeliverabilityTestReport

Retrieve the results of a predictive inbox placement test.

GetDomainDeliverabilityCampaign

Retrieve all the deliverability data for a specific campaign. This data is available

GetDomainStatisticsReport

Retrieve inbox placement and engagement rates for the domains that you use to send

GetEmailIdentity

Provides information about a specific identity associated with your Amazon Pinpoint account,

ListConfigurationSets

List all of the configuration sets associated with your Amazon Pinpoint account in the current

ListDedicatedIpPools

List all of the dedicated IP pools that exist in your Amazon Pinpoint account in the current

ListDeliverabilityTestReports

Show a list of the predictive inbox placement tests that you've performed, regardless of their statuses. For

ListDomainDeliverabilityCampaigns

Retrieve deliverability data for all the campaigns that used a specific domain to send

ListEmailIdentities

Returns a list of all of the email identities that are associated with your Amazon Pinpoint

ListTagsForResource

Retrieve a list of the tags (keys and values) that are associated with a specified

PutAccountDedicatedIpWarmupAttributes

Enable or disable the automatic warm-up feature for dedicated IP addresses.

PutAccountSendingAttributes

Enable or disable the ability of your account to send email.

PutConfigurationSetDeliveryOptions

Associate a configuration set with a dedicated IP pool. You can use dedicated IP pools

PutConfigurationSetReputationOptions

Enable or disable collection of reputation metrics for emails that you send using a

PutConfigurationSetSendingOptions

Enable or disable email sending for messages that use a particular configuration set

PutConfigurationSetTrackingOptions

Specify a custom domain to use for open and click tracking elements in email that you

PutDedicatedIpInPool

Move a dedicated IP address to an existing dedicated IP pool.

PutDedicatedIpWarmupAttributes

PutDeliverabilityDashboardOption

Enable or disable the Deliverability dashboard for your Amazon Pinpoint account. When you enable the

PutEmailIdentityDkimAttributes

Used to enable or disable DKIM authentication for an email identity.

PutEmailIdentityFeedbackAttributes

Used to enable or disable feedback forwarding for an identity. This setting determines

PutEmailIdentityMailFromAttributes

Used to enable or disable the custom Mail-From domain configuration for an email

SendEmail

Sends an email message. You can use the Amazon Pinpoint Email API to send two types of

TagResource

Add one or more tags (keys and values) to a specified resource. A

UntagResource

Remove one or more tags (keys and values) from a specified resource.

UpdateConfigurationSetEventDestination

Update the configuration of an event destination for a configuration set.