Module aws_sdk_pinpointemail::output

Modules

create_configuration_set_event_destination_output

See CreateConfigurationSetEventDestinationOutput

create_configuration_set_output

See CreateConfigurationSetOutput

create_dedicated_ip_pool_output

See CreateDedicatedIpPoolOutput

create_deliverability_test_report_output

See CreateDeliverabilityTestReportOutput

create_email_identity_output

See CreateEmailIdentityOutput

delete_configuration_set_event_destination_output

See DeleteConfigurationSetEventDestinationOutput

delete_configuration_set_output

See DeleteConfigurationSetOutput

delete_dedicated_ip_pool_output

See DeleteDedicatedIpPoolOutput

delete_email_identity_output

See DeleteEmailIdentityOutput

get_account_output

See GetAccountOutput

get_blacklist_reports_output

See GetBlacklistReportsOutput

get_configuration_set_event_destinations_output

See GetConfigurationSetEventDestinationsOutput

get_configuration_set_output

See GetConfigurationSetOutput

get_dedicated_ip_output

See GetDedicatedIpOutput

get_dedicated_ips_output

See GetDedicatedIpsOutput

get_deliverability_dashboard_options_output

See GetDeliverabilityDashboardOptionsOutput

get_deliverability_test_report_output

See GetDeliverabilityTestReportOutput

get_domain_deliverability_campaign_output

See GetDomainDeliverabilityCampaignOutput

get_domain_statistics_report_output

See GetDomainStatisticsReportOutput

get_email_identity_output

See GetEmailIdentityOutput

list_configuration_sets_output

See ListConfigurationSetsOutput

list_dedicated_ip_pools_output

See ListDedicatedIpPoolsOutput

list_deliverability_test_reports_output

See ListDeliverabilityTestReportsOutput

list_domain_deliverability_campaigns_output

See ListDomainDeliverabilityCampaignsOutput

list_email_identities_output

See ListEmailIdentitiesOutput

list_tags_for_resource_output

See ListTagsForResourceOutput

put_account_dedicated_ip_warmup_attributes_output

See PutAccountDedicatedIpWarmupAttributesOutput

put_account_sending_attributes_output

See PutAccountSendingAttributesOutput

put_configuration_set_delivery_options_output

See PutConfigurationSetDeliveryOptionsOutput

put_configuration_set_reputation_options_output

See PutConfigurationSetReputationOptionsOutput

put_configuration_set_sending_options_output

See PutConfigurationSetSendingOptionsOutput

put_configuration_set_tracking_options_output

See PutConfigurationSetTrackingOptionsOutput

put_dedicated_ip_in_pool_output

See PutDedicatedIpInPoolOutput

put_dedicated_ip_warmup_attributes_output

See PutDedicatedIpWarmupAttributesOutput

put_deliverability_dashboard_option_output

See PutDeliverabilityDashboardOptionOutput

put_email_identity_dkim_attributes_output

See PutEmailIdentityDkimAttributesOutput

put_email_identity_feedback_attributes_output

See PutEmailIdentityFeedbackAttributesOutput

put_email_identity_mail_from_attributes_output

See PutEmailIdentityMailFromAttributesOutput

send_email_output

See SendEmailOutput

tag_resource_output

See TagResourceOutput

untag_resource_output

See UntagResourceOutput

update_configuration_set_event_destination_output

See UpdateConfigurationSetEventDestinationOutput

Structs

CreateConfigurationSetEventDestinationOutput

An HTTP 200 response if the request succeeds, or an error message if the request

CreateConfigurationSetOutput

An HTTP 200 response if the request succeeds, or an error message if the request

CreateDedicatedIpPoolOutput

An HTTP 200 response if the request succeeds, or an error message if the request

CreateDeliverabilityTestReportOutput

Information about the predictive inbox placement test that you created.

CreateEmailIdentityOutput

If the email identity is a domain, this object contains tokens that you can use to

DeleteConfigurationSetEventDestinationOutput

An HTTP 200 response if the request succeeds, or an error message if the request

DeleteConfigurationSetOutput

An HTTP 200 response if the request succeeds, or an error message if the request

DeleteDedicatedIpPoolOutput

An HTTP 200 response if the request succeeds, or an error message if the request

DeleteEmailIdentityOutput

An HTTP 200 response if the request succeeds, or an error message if the request

GetAccountOutput

A list of details about the email-sending capabilities of your Amazon Pinpoint account in the

GetBlacklistReportsOutput

An object that contains information about blacklist events.

GetConfigurationSetEventDestinationsOutput

Information about an event destination for a configuration set.

GetConfigurationSetOutput

Information about a configuration set.

GetDedicatedIpOutput

Information about a dedicated IP address.

GetDedicatedIpsOutput

Information about the dedicated IP addresses that are associated with your Amazon Pinpoint

GetDeliverabilityDashboardOptionsOutput

An object that shows the status of the Deliverability dashboard for your Amazon Pinpoint account.

GetDeliverabilityTestReportOutput

The results of the predictive inbox placement test.

GetDomainDeliverabilityCampaignOutput

An object that contains all the deliverability data for a specific campaign. This data

GetDomainStatisticsReportOutput

An object that includes statistics that are related to the domain that you

GetEmailIdentityOutput

Details about an email identity.

ListConfigurationSetsOutput

A list of configuration sets in your Amazon Pinpoint account in the current AWS Region.

ListDedicatedIpPoolsOutput

A list of dedicated IP pools.

ListDeliverabilityTestReportsOutput

A list of the predictive inbox placement test reports that are available for your account, regardless of

ListDomainDeliverabilityCampaignsOutput

An array of objects that provide deliverability data for all the campaigns that used a

ListEmailIdentitiesOutput

A list of all of the identities that you've attempted to verify for use with Amazon Pinpoint,

ListTagsForResourceOutput

PutAccountDedicatedIpWarmupAttributesOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutAccountSendingAttributesOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutConfigurationSetDeliveryOptionsOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutConfigurationSetReputationOptionsOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutConfigurationSetSendingOptionsOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutConfigurationSetTrackingOptionsOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutDedicatedIpInPoolOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutDedicatedIpWarmupAttributesOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutDeliverabilityDashboardOptionOutput

A response that indicates whether the Deliverability dashboard is enabled for your Amazon Pinpoint

PutEmailIdentityDkimAttributesOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutEmailIdentityFeedbackAttributesOutput

An HTTP 200 response if the request succeeds, or an error message if the request

PutEmailIdentityMailFromAttributesOutput

An HTTP 200 response if the request succeeds, or an error message if the request

SendEmailOutput

A unique message ID that you receive when Amazon Pinpoint accepts an email for sending.

TagResourceOutput

UntagResourceOutput

UpdateConfigurationSetEventDestinationOutput

An HTTP 200 response if the request succeeds, or an error message if the request